resolve filename relative to search path
PR #3 changes how the filepath is composed. Unfortunately if the filename is not
found in the root it searches for the filename itself, relative to the current
working directory.

When using find-up to find a file in a disjoint directory, where that file also
exists in the current working directory (say a `package.json` file), the local
file is returned.

This commit adds tests for find-up returning `null` if a file is not found, and
ensures the filename is resolved relative to the root and not the working
directory.